fix(AttendanceOperator): 입력 예외 발생 시 무한 입력 받도록 수정
