--- src/main/java/controller/EditionOperator.java
@@ -1,12 +1,0 @@
-package controller;
-
-import domain.AttendanceBook;
-import java.time.LocalDate;
-
-public class EditionOperator implements OptionOperator {
-
-    @Override
-    public void process(AttendanceBook attendanceBook, LocalDate attendanceDate) {
-
-    }
-}
